pyproject.toml
[project]
name = "snipara-fastapi"
version = "1.9.0"
description = "Hosted MCP server for RLM SaaS - Context-efficient documentation queries"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "RLM SaaS Team" }]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "redis>=5.0.0",
    "prisma>=0.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "bcrypt>=4.1.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pdoc>=15.0.0",
]
benchmark = [
    "anthropic>=0.40.0",
    "tiktoken>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[project.scripts]
snipara-benchmark = "benchmarks.runner:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "benchmarks"]

[tool.mypy]
python_version = "3.11"
strict = true
